pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mela-parser"
version = "0.1.0"
description = "Extract recipes from EPUB cookbooks to Mela format"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "Mela Parser Contributors"}
]
keywords = ["recipe", "epub", "parser", "mela", "cookbook"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "beautifulsoup4>=4.13.4",
    "ebooklib>=0.18",
    "html2text>=2024.2.26",
    "markitdown[all]>=0.1.3",
    "openai>=1.66.3",
    "pillow>=11.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.11.6",
]

[project.scripts]
mela-parse = "mela_parser.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mela_parser"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py313"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E501"]  # Allow longer lines in tests
